+++ src/main/java/gov/cida/cdat/io/container/FileStreamContainer.java
@@ -1,4 +1,4 @@
-package gov.cida.cdat.io.stream;
+package gov.cida.cdat.io.container;
 
 import gov.cida.cdat.exception.StreamInitException;
