+++ src/test/java/gov/cida/cdat/control/SCManagerControlInterruptTest.java
@@ -6,9 +6,9 @@ import gov.cida.cdat.control.Control;
 import gov.cida.cdat.control.SCManager;
 import gov.cida.cdat.exception.CdatException;
 import gov.cida.cdat.exception.StreamInitException;
-import gov.cida.cdat.io.stream.DataPipe;
-import gov.cida.cdat.io.stream.SimpleStreamContainer;
-import gov.cida.cdat.io.stream.StreamContainer;
+import gov.cida.cdat.io.container.DataPipe;
+import gov.cida.cdat.io.container.SimpleStreamContainer;
+import gov.cida.cdat.io.container.StreamContainer;
 import gov.cida.cdat.message.Message;
 import gov.cida.cdat.service.PipeWorker;
 import gov.cida.cdat.service.Worker;
