+++ src/test/java/gov/cida/cdat/control/SCManagerControlSuccessTest.java
@@ -5,8 +5,8 @@ import gov.cida.cdat.TestUtils;
 import gov.cida.cdat.control.Callback;
 import gov.cida.cdat.control.Control;
 import gov.cida.cdat.control.SCManager;
-import gov.cida.cdat.io.stream.DataPipe;
-import gov.cida.cdat.io.stream.SimpleStreamContainer;
+import gov.cida.cdat.io.container.DataPipe;
+import gov.cida.cdat.io.container.SimpleStreamContainer;
 import gov.cida.cdat.message.Message;
 import gov.cida.cdat.service.PipeWorker;
 import gov.cida.cdat.service.Worker;
