+++ src/test/java/gov/cida/cdat/db/DBTesting.java
@@ -6,8 +6,8 @@ import gov.cida.cdat.control.Control;
 import gov.cida.cdat.control.SCManager;
 import gov.cida.cdat.exception.CdatException;
 import gov.cida.cdat.io.Closer;
-import gov.cida.cdat.io.stream.SimpleStreamContainer;
-import gov.cida.cdat.io.stream.TransformStreamContainer;
+import gov.cida.cdat.io.container.SimpleStreamContainer;
+import gov.cida.cdat.io.container.TransformStreamContainer;
 import gov.cida.cdat.message.Message;
 import gov.cida.cdat.service.Worker;
 import gov.cida.cdat.transform.Transformer;
